+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 - Bumped AndroidX media3 to `1.4.1` from `1.3.1`.
 
+### 🛠️ Fixes
+
+- Ensure "year" field returns a year.
+
 ## [0.3.0] - 2024-09-01
 
 ### ⚡ Changes